Soft delete en todas las entidades, usar dtos

+++ src/main/java/com/vncarca/arcasys/persona/services/PersonaServiceImp.java
@@ -2,7 +2,12 @@ package com.vncarca.arcasys.persona.services;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
+import com.vncarca.arcasys.persona.model.PersonaDtoExtends;
+import org.aspectj.weaver.ast.Instanceof;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -18,50 +23,49 @@ import com.vncarca.arcasys.persona.repository.PersonaRepository;
 public class PersonaServiceImp implements PersonaService {
 	@Autowired
 	PersonaRepository personaRepository;
-
+  @Autowired
+	ModelMapper modelMapper;
 	@Override
-	public Page<Persona> findAll(Pageable pageable) {
+	public Page<PersonaDtoExtends> findAll(Pageable pageable) {
 
-		return personaRepository.findAll(pageable);
+		return personaRepository.findAll(pageable).map(this::convertToDto);
 	}
 
 	@Override
-	public List<Persona> findAll() {
-
-		return personaRepository.findAll();
+	public List<PersonaDtoExtends> findAll() {
+		return personaRepository.findAll().stream().map(this::convertToDto).collect(Collectors.toList());
 	}
 
 	@Override
-	public Persona save(Persona persona) {
-
-		return personaRepository.save(persona);
+	public PersonaDtoExtends save(PersonaDtoExtends persona) {
+		return convertToDto(personaRepository.save(convertToEntity(persona)));
 	}
 
 	@Override
-	public Persona findById(Long id) {
-
-		return personaRepository.findById(id).orElseThrow(() -> new NoSuchElementException(
+	public PersonaDtoExtends findById(Long id) {
+		return personaRepository.findById(id).map(this::convertToDto).orElseThrow(() -> new NoSuchElementException(
 				"Persona con ID: " + id.toString() + " no existe en el servidor"));
 	}
 
 	@Override
 	public void delete(Long id) {
-		personaRepository.deleteById(id);
+		if(Objects.equals(findById(id).getId(), id))
+			personaRepository.deleteById(id);
 	}
 
 	@Override
-	public Page<Persona> findByCedula(Pageable pageable, String cedula) {
-		return personaRepository.findByCedula(pageable, cedula);
+	public Page<PersonaDtoExtends> findByCedula(Pageable pageable, String cedula) {
+		return personaRepository.findByCedula(pageable, cedula).map(this::convertToDto);
 	}
 
 	@Override
-	public PersonaDto convertToDto(Persona entity) {
-		return null;
+	public PersonaDtoExtends convertToDto(Persona entity) {
+		return modelMapper.map(entity, PersonaDtoExtends.class);
 	}
 
 	@Override
-	public Persona convertToEntity(PersonaDto dto) {
-		// TODO:
-		return null;
+	public Persona convertToEntity(PersonaDtoExtends dto) {
+		return modelMapper.map(dto,Persona.class);
 	}
+
 }
